Completed refactoring
