+++ docs/auth/datalab/siesta/BusinessLogic/ExtractPairs/ExtractPairsSimple$.html
@@ -3,9 +3,9 @@
         <head>
           <meta http-equiv="X-UA-Compatible" content="IE=edge" />
           <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
-          <title>SIESTA preprocess documentation  - auth.datalab.siesta.BusinessLogic.ExtractPairs.ExtractPairsSimple</title>
-          <meta name="description" content="SIESTA preprocess documentation - auth.datalab.siesta.BusinessLogic.ExtractPairs.ExtractPairsSimple" />
-          <meta name="keywords" content="SIESTA preprocess documentation auth.datalab.siesta.BusinessLogic.ExtractPairs.ExtractPairsSimple" />
+          <title>SIESTA preprocess documentation  - auth.datalab.siesta.BusinessLogic.ExtractPairs.ExtractPairs</title>
+          <meta name="description" content="SIESTA preprocess documentation - auth.datalab.siesta.BusinessLogic.ExtractPairs.ExtractPairs" />
+          <meta name="keywords" content="SIESTA preprocess documentation auth.datalab.siesta.BusinessLogic.ExtractPairs.ExtractPairs" />
           <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
